feat: basic test
