+++ .github/workflows/ci.yml
@@ -6,6 +6,10 @@ jobs:
   test:
     runs-on: ubuntu-latest
 
+    env:
+      HUMANITEC_ORG: ${{ secrets.HUMANITEC_ORG }}
+      HUMANITEC_TOKEN: ${{ secrets.HUMANITEC_TOKEN }}
+
     strategy:
       matrix:
         node-version: [16.x, 18.x]
@@ -18,4 +22,5 @@ jobs:
           node-version: ${{ matrix.node-version }}
       - run: npm ci
       - run: npm run lint
+      - run: npm test
       - run: npm run package
