+++ fixtures/Dockerfile
@@ -1,0 +1,3 @@
+FROM scratch
+
+ENV almost empty
