+++ package.json
@@ -5,11 +5,13 @@
   "main": "dist/index.js",
   "scripts": {
     "lint": "eslint index.ts docker.ts humanitec.ts",
-    "package": "ncc build index.ts -o dist"
+    "package": "ncc build index.ts -o dist",
+    "test": "jest"
   },
   "author": "chris.stephenson@humanitec.com",
   "license": "ISC",
   "devDependencies": {
+    "@jest/globals": "^29.3.1",
     "@tsconfig/node16": "^1.0.3",
     "@types/node": "^16.18.3",
     "@types/node-fetch": "^2.6.2",
@@ -19,6 +21,8 @@
     "eslint": "^8.28.0",
     "eslint-config-eslint": "^7.0.0",
     "eslint-config-google": "^0.14.0",
+    "jest": "^29.3.1",
+    "ts-jest": "^29.0.3",
     "typescript": "^4.9.3"
   },
   "dependencies": {
